Data.Convertible

Description

This is a library to provide a uniform interface for safe conversions between different types of data. To get started reading about it, consult:

Data.Convertible.Base for information about the conversions themselves

Data.Convertible.Utils for helpful tools for people writing Convertible instances

Data.Convertible.Instances for a large collection of ready-built Convertible instances

You can import these modules individually, or this module will export the entire library for you.

Documentation

module Data.Convertible.Base

module Data.Convertible.Utils