Data.Convertible.Instances.Text

Description

Text instances for Convertible.

Copyright (C) 2011 MailRank, Inc. bos@mailrank.com

All rights reserved.

For license and copyright information, see the file COPYRIGHT

These instances perform conversion between text-like types such as Text, ByteString, and the like.

The instances do not include conversions between ByteString and Text or String, since such conversions cannot safely be performed without knowing the encoding of the ByteString.

Orphan instances

Convertible Char Builder

Methods

safeConvert :: Char -> ConvertResult Builder

Convertible Char Text

Methods

safeConvert :: Char -> ConvertResult Text

Convertible Char Text

Methods

safeConvert :: Char -> ConvertResult Text

Convertible Word8 ByteString

Methods

safeConvert :: Word8 -> ConvertResult ByteString

Convertible Word8 ByteString

Methods

safeConvert :: Word8 -> ConvertResult ByteString

Convertible Word8 Builder

Methods

safeConvert :: Word8 -> ConvertResult Builder

Convertible ByteString ByteString

Methods

safeConvert :: ByteString -> ConvertResult ByteString

Convertible ByteString Builder

Methods

safeConvert :: ByteString -> ConvertResult Builder

Convertible ByteString Builder

Methods

safeConvert :: ByteString -> ConvertResult Builder

Convertible ByteString Text

Methods

safeConvert :: ByteString -> ConvertResult Text

Convertible ByteString Text

Methods

safeConvert :: ByteString -> ConvertResult Text

Convertible ByteString ByteString

Methods

safeConvert :: ByteString -> ConvertResult ByteString

Convertible ByteString Builder

Methods

safeConvert :: ByteString -> ConvertResult Builder

Convertible ByteString Builder

Methods

safeConvert :: ByteString -> ConvertResult Builder

Convertible ByteString Text

Methods

safeConvert :: ByteString -> ConvertResult Text

Convertible ByteString Text

Methods

safeConvert :: ByteString -> ConvertResult Text

Convertible Builder ByteString

Methods

safeConvert :: Builder -> ConvertResult ByteString

Convertible Builder ByteString

Methods

safeConvert :: Builder -> ConvertResult ByteString

Convertible Builder Builder

Methods

safeConvert :: Builder -> ConvertResult Builder

Convertible Builder Text

Methods

safeConvert :: Builder -> ConvertResult Text

Convertible Builder Text

Methods

safeConvert :: Builder -> ConvertResult Text

Convertible Text ByteString

Methods

safeConvert :: Text -> ConvertResult ByteString

Convertible Text ByteString

Methods

safeConvert :: Text -> ConvertResult ByteString

Convertible Text Builder

Methods

safeConvert :: Text -> ConvertResult Builder

Convertible Text Builder

Methods

safeConvert :: Text -> ConvertResult Builder

Convertible Text Text

Methods

safeConvert :: Text -> ConvertResult Text

Convertible Text ByteString

Methods

safeConvert :: Text -> ConvertResult ByteString

Convertible Text ByteString

Methods

safeConvert :: Text -> ConvertResult ByteString

Convertible Text Builder

Methods

safeConvert :: Text -> ConvertResult Builder

Convertible Text Builder

Methods

safeConvert :: Text -> ConvertResult Builder

Convertible Text Text

Methods

safeConvert :: Text -> ConvertResult Text

Convertible ByteString [Word8]

Methods

safeConvert :: ByteString -> ConvertResult [Word8]

Convertible ByteString [Word8]

Methods

safeConvert :: ByteString -> ConvertResult [Word8]

Convertible Builder [Char]

Methods

safeConvert :: Builder -> ConvertResult [Char]

Convertible Text [Char]

Methods

safeConvert :: Text -> ConvertResult [Char]

Convertible Text [Char]

Methods

safeConvert :: Text -> ConvertResult [Char]

Convertible [Char] Builder

Methods

safeConvert :: [Char] -> ConvertResult Builder

Convertible [Char] Text

Methods

safeConvert :: [Char] -> ConvertResult Text

Convertible [Char] Text

Methods

safeConvert :: [Char] -> ConvertResult Text

Convertible [Word8] ByteString

Methods

safeConvert :: [Word8] -> ConvertResult ByteString

Convertible [Word8] ByteString

Methods

safeConvert :: [Word8] -> ConvertResult ByteString

Convertible [Word8] Builder

Methods

safeConvert :: [Word8] -> ConvertResult Builder