Data.Convertible.Instances.Time

Description

Instances to convert between various time structures, both old- and new-style.

At present, this module does not do full input validation. That is, it is possible to get an exception rather than a Left result from these functions if your input is invalid, particularly when converting from the old-style System.Time structures.

Copyright (C) 2009-2011 John Goerzen jgoerzen@complete.org

All rights reserved.

For license and copyright information, see the file LICENSE

Orphan instances

Convertible Double CTime

Methods

safeConvert :: Double -> ConvertResult CTime

Convertible Double TimeDiff

Methods

safeConvert :: Double -> ConvertResult TimeDiff

Convertible Double POSIXTime

Methods

safeConvert :: Double -> ConvertResult POSIXTime

Convertible Double UTCTime

Methods

safeConvert :: Double -> ConvertResult UTCTime

Convertible Int CTime

Methods

safeConvert :: Int -> ConvertResult CTime

Convertible Int POSIXTime

Methods

safeConvert :: Int -> ConvertResult POSIXTime

Convertible Int UTCTime

Methods

safeConvert :: Int -> ConvertResult UTCTime

Convertible Integer CTime

Methods

safeConvert :: Integer -> ConvertResult CTime

Convertible Integer ClockTime

Methods

safeConvert :: Integer -> ConvertResult ClockTime

Convertible Integer TimeDiff

Methods

safeConvert :: Integer -> ConvertResult TimeDiff

Convertible Integer POSIXTime

Methods

safeConvert :: Integer -> ConvertResult POSIXTime

Convertible Integer UTCTime

Methods

safeConvert :: Integer -> ConvertResult UTCTime

Convertible Rational POSIXTime

Methods

safeConvert :: Rational -> ConvertResult POSIXTime

Convertible Rational UTCTime

Methods

safeConvert :: Rational -> ConvertResult UTCTime

Convertible CTime Double

Methods

safeConvert :: CTime -> ConvertResult Double

Convertible CTime Int

Methods

safeConvert :: CTime -> ConvertResult Int

Convertible CTime Integer

Methods

safeConvert :: CTime -> ConvertResult Integer

Convertible CTime ClockTime

Methods

safeConvert :: CTime -> ConvertResult ClockTime

Convertible CTime CalendarTime

Methods

safeConvert :: CTime -> ConvertResult CalendarTime

Convertible CTime ZonedTime

Methods

safeConvert :: CTime -> ConvertResult ZonedTime

Convertible CTime POSIXTime

Methods

safeConvert :: CTime -> ConvertResult POSIXTime

Convertible CTime UTCTime

Methods

safeConvert :: CTime -> ConvertResult UTCTime

Convertible ClockTime Integer

Methods

safeConvert :: ClockTime -> ConvertResult Integer

Convertible ClockTime CTime

Methods

safeConvert :: ClockTime -> ConvertResult CTime

Convertible ClockTime CalendarTime

Methods

safeConvert :: ClockTime -> ConvertResult CalendarTime

Convertible ClockTime ZonedTime

Methods

safeConvert :: ClockTime -> ConvertResult ZonedTime

Convertible ClockTime POSIXTime

Methods

safeConvert :: ClockTime -> ConvertResult POSIXTime

Convertible ClockTime UTCTime

Methods

safeConvert :: ClockTime -> ConvertResult UTCTime

Convertible CalendarTime CTime

Methods

safeConvert :: CalendarTime -> ConvertResult CTime

Convertible CalendarTime ClockTime

Methods

safeConvert :: CalendarTime -> ConvertResult ClockTime

Convertible CalendarTime ZonedTime

Methods

safeConvert :: CalendarTime -> ConvertResult ZonedTime

Convertible CalendarTime POSIXTime

Methods

safeConvert :: CalendarTime -> ConvertResult POSIXTime

Convertible CalendarTime UTCTime

Methods

safeConvert :: CalendarTime -> ConvertResult UTCTime

Convertible TimeDiff Double

Methods

safeConvert :: TimeDiff -> ConvertResult Double

Convertible TimeDiff Integer

Methods

safeConvert :: TimeDiff -> ConvertResult Integer

Convertible TimeDiff Rational

Methods

safeConvert :: TimeDiff -> ConvertResult Rational

Convertible TimeDiff NominalDiffTime

Methods

safeConvert :: TimeDiff -> ConvertResult NominalDiffTime

Convertible ZonedTime CTime

Methods

safeConvert :: ZonedTime -> ConvertResult CTime

Convertible ZonedTime ClockTime

Methods

safeConvert :: ZonedTime -> ConvertResult ClockTime

Convertible ZonedTime CalendarTime

Methods

safeConvert :: ZonedTime -> ConvertResult CalendarTime

Convertible ZonedTime POSIXTime

Methods

safeConvert :: ZonedTime -> ConvertResult POSIXTime

Convertible ZonedTime UTCTime

Methods

safeConvert :: ZonedTime -> ConvertResult UTCTime

Convertible POSIXTime Double

Methods

safeConvert :: POSIXTime -> ConvertResult Double

Convertible POSIXTime Int

Methods

safeConvert :: POSIXTime -> ConvertResult Int

Convertible POSIXTime Integer

Methods

safeConvert :: POSIXTime -> ConvertResult Integer

Convertible POSIXTime Rational

Methods

safeConvert :: POSIXTime -> ConvertResult Rational

Convertible POSIXTime CTime

Methods

safeConvert :: POSIXTime -> ConvertResult CTime

Convertible POSIXTime ClockTime

Methods

safeConvert :: POSIXTime -> ConvertResult ClockTime

Convertible POSIXTime CalendarTime

Methods

safeConvert :: POSIXTime -> ConvertResult CalendarTime

Convertible POSIXTime ZonedTime

Methods

safeConvert :: POSIXTime -> ConvertResult ZonedTime

Convertible POSIXTime UTCTime

Methods

safeConvert :: POSIXTime -> ConvertResult UTCTime

Convertible UTCTime Double

Methods

safeConvert :: UTCTime -> ConvertResult Double

Convertible UTCTime Int

Methods

safeConvert :: UTCTime -> ConvertResult Int

Convertible UTCTime Integer

Methods

safeConvert :: UTCTime -> ConvertResult Integer

Convertible UTCTime Rational

Methods

safeConvert :: UTCTime -> ConvertResult Rational

Convertible UTCTime CTime

Methods

safeConvert :: UTCTime -> ConvertResult CTime

Convertible UTCTime ClockTime

Methods

safeConvert :: UTCTime -> ConvertResult ClockTime

Convertible UTCTime CalendarTime

Methods

safeConvert :: UTCTime -> ConvertResult CalendarTime

Convertible UTCTime ZonedTime

Methods

safeConvert :: UTCTime -> ConvertResult ZonedTime

Convertible UTCTime POSIXTime

Methods

safeConvert :: UTCTime -> ConvertResult POSIXTime

Convertible NominalDiffTime TimeDiff

Methods

safeConvert :: NominalDiffTime -> ConvertResult TimeDiff